lv2vst: Fix crash when switching to generic UI on Windows
On Windows, LV2 UI can be shown without error. But when switching to
generic UI, host will crash. GDB cannot print a proper backtrace, but
shows a stack warning (seems to be an internal warning in MSVC runtime).

Then I let the VST dispacher function print what opcode executes, then
it crashed at opcode = 8 (effGetParamName). Here executes
get_parameter_name(), and it uses strncpyn() to set param name on host.
On REAPER, it assumed that the maximum string length is 256, however
this will overflow on Windows (but OK on Linux).

So solution is to set to 255. Everything is OK.